Test harness, Implement better data loading, add UBI, lots of cleanup and bug fixes (#217)
* Fixes ch5.1 with another virtual environment

* Cleaned up requirements file

* Adds dependencies to ch5 venv to allow entire notebook to run on that kernel

* - Creates aips.indexer to handle data operations for all datasets/collections
  - idempotent building of collections
  - standardized github interations: cloning, extracting
  - all github repositories download to "data/repositories/"
- Integrates aips.indexer into all chapters
- Implemented Collection.get_document_count, Engine.does_collection_extist, Engine.is_collection_healthy
- Moves "has_promotion" product data logic to data loader.
- Unjankifies (Un-jank-ify, technical term) from_csv, spliting out logic around "category" and id

* ok

* Unstubs spladerunner using 0.1.6, fixing 7.14

* Runs 7.14 to have output, adds missing deps to ch5 requirements, updates ch5.1 to have correct kernel selected by default

* Updates the ch5.1 kernel name

* Logs github progress with tqdm, formats and documents indexer.py, re-verifies code

* Updates indexer documentation

* Fixes typo in indexer collection dictionary

* adds test suite, fixes import

* Several data feature namespace fixes

* Adds retry logic magic function, fixes other ch11 import

* Increases solr memory to 1g, fixing various crashes throughout the notebooks.
Removes retry logic as remedied by the memory increase
Corrects typo'd movie dataset values

* ch13 Fixes tokenizer route
ch12 Fixes promotion index name and config
Removes unused retry code
Splits out test suite into cells and

* Fixes ch12, ch13, ch14 and ch15
Integrates standardized spark session generation

* Fixes ch14 - correct eval_strategy arg name, updates accelerate version to 1.9.0
Fixes ch5/ch6 skg data

* Adds opensearch ubi
Adjusts opensearch schema
Fixing query_string field bug

* Corrects loading of ndjson data for dashboard uploading

* Fixes Chapter 8 index time boosting field name

* Fixes ltr opensearch chapters
Improves test logic to test all engines, cleans up code a bit
Correct 10.4 list of documents
Corrects logging bug in chapter 12 LTR evaluate_model
Corrects opensearch schema config for products_with_promotions and upc as keyword
Cleans OpenSearchEngine

* upgrades to opensearch 2.18
attempts to fix 15.14

* Fixes a few missing idempotent data loading spots
Corrects schema for chapter 15 for title matching
OpenSearch LTR no longer re-creates the ltr store if already enabled
AIPS OpenSearch id override now is properly handled to retrieve _id from metadata for spark loading.
Cleans errored notebooks
Test harness cleaned

* Test harness with all notebooks + all engines passing

* Fixes ch11.4 functionality but does not refactor to LTR abstraction. Tested C11 notebooks

* Ubi: Adds signaltime to ubi event ingestion example
        selects only necessary data from ubi_queries when batch importing into AIPS signals collection

* Updates UBI notebook infomation and logic to be up to date with 2.18 practices.

* cleans print statements, reruns ch3 with solr

* Assigns collection being returned from .build_collection() throughout the code base to eliminate printing of var at end of cells.
Caches downloaded packages during ch5_spacy building
Fixes job from_csv indexing (and intended jobs count) with multiline csv reading (ch 5.3)
Ch6.19 corrects tokenize=True (forgotten change, it's correct in the manuscript)
Fixes stackexchange building for OS, correctly using append instead of overwrite
Adds append logic to solr write (it doesn't seem to be used in the dataframe writer anyway when writing w/ solr plugin?)
Makes 10.4 idempotent

* ch10 - Functional and run with solr, however Inconsistencies still (perhaps tmdb related).
ch11 - formatted properly (at least 120char width), finished
ch12 - Reran for solr, mostly accurate with minor inconsistencies
ch13 - Everything correct except for ranking of 13.2 where the default order changed?
ch14 - Rerun for Solr to fix errored listined, everything accurate.
ch15 - Functional, run with solr. Rankings are the same, except for some of the figure listings which have a slightly different rank. The vector scores and ranking throughout 15.2 are accurate and in sync.

* Implements Engine.AdvancedFeatures first used for dual indexing. This feature will be expanded in the weaviate/embedded solr integration.
Cleans up sloppy dual indexing code
Cleans up some white space and adds some pythonic null coalescing

* Notebook cleanup

* Listing 8.1 cleanup (like book)

* UBI Notebook updates

* Fix opensearch version and update UBI notebook to skip live simulation by default

---------

+++ engines/opensearch/OpenSearchEngine.py
@@ -2,7 +2,7 @@ import json
 
 import requests
 
-from engines.Engine import Engine
+from engines.Engine import AdvancedFeatures, Engine
 from engines.opensearch.config import OPENSEARCH_URL, SCHEMAS
 from engines.opensearch.OpenSearchCollection import OpenSearchCollection
 
@@ -12,29 +12,39 @@ class OpenSearchEngine(Engine):
     def __init__(self):
         super().__init__("OpenSearch")
 
+    def get_supported_advanced_features(self):
+        return [AdvancedFeatures.LTR]
+
     def health_check(self):
         status = requests.get(STATUS_URL).json()["status"] in ["green", "yellow"]
-        if status:
-            print("OpenSearch engine is online")
+        if status: print("OpenSearch engine is online")
         return status
     
+    def does_collection_exist(self, name):
+        response = requests.get(f"{STATUS_URL}/{name}")
+        return response.status_code == 200
+    
+    def is_collection_healthy(self, name, expected_count, log=False):
+        exists = self.does_collection_exist(name)
+        document_count = self.get_collection(name).get_document_count()
+        if log: print(exists, document_count)
+        return exists and document_count  == expected_count
+
     def print_status(self, response):
-        #print(json.dumps(response, indent=2))
-        "Prints the resulting status of a search engine request"
+        print(json.dumps(response, indent=2))
         pass
 
-    def create_collection(self, name, log=False):
-        print(f'Wiping "{name}" collection')
-        response = requests.delete(f"{OPENSEARCH_URL}/{name}").json()
+    def create_collection(self, name, force_rebuild=True, log=False):
+        if force_rebuild:
+            print(f'Wiping "{name}" collection')
+            response = requests.delete(f"{OPENSEARCH_URL}/{name}").json()
 
         print(f'Creating "{name}" collection')
         collection = self.get_collection(name)
         request = SCHEMAS[name]["schema"] if name in SCHEMAS else {}
         response = requests.put(f"{OPENSEARCH_URL}/{name}", json=request).json()
-        if log: 
-            print("Schema:", json.dumps(request, indent=2))
-        if log: 
-            print("Status:", json.dumps(response, indent=2))
+        if log: print("Schema:", json.dumps(request, indent=2))
+        if log: print("Status:", json.dumps(response, indent=2))
         return collection
 
     def get_collection(self, name):
